Rename the folder home to splash
